# docs/security/openapi-sso-s2s.yaml
...
openapi: 3.1.0
info:
  title: DreamSSO service-to-service API (mTLS / API Shield)
  version: "2026-07-13"
  description: >
    Server-to-server endpoints on sso.dreamxwarden.ca that sit behind the CF
    mTLS require-cert set: OIDC token/userinfo, the RP event channel, and the
    account-portal /internal/* family. Client mTLS is terminated and verified at
    the Cloudflare edge (cf.tls_client_auth.cert_verified); origins never see the
    cert. App-layer auth is private_key_jwt (RFC 7523), Bearer access token, or a
    signed event JWT as noted per operation. /account/* resource routes are in the
    require-cert set but out of scope for schema validation here.
servers:
  - url: https://sso.dreamxwarden.ca
security:
  - edgeMutualTLS: []
...
components:
  securitySchemes:
    edgeMutualTLS:
      type: mutualTLS
      description: >
        Client certificate presented by the calling backend and verified at the
        Cloudflare edge. Callers use a CF-managed client cert (videosite) or a
        homelab-CA cert uploaded via BYOCA (account BFF, once provisioned).
    privateKeyJwt:
      type: apiKey
      in: cookie
      name: client_assertion
      description: >
        RFC 7523 private_key_jwt in the form body (not really a cookie; modeled as
        apiKey for tooling). client_assertion_type + client_assertion fields; iss=sub=client_id,
        aud in [issuer, issuer/token], signed by the client's registered key (jwks_uri/jwks).
    bearerAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: SSO access token, aud=issuer, carries space-delimited scope.
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        error_description: { type: string }
      required: [error]
    ClientAssertion:
      type: object
      properties:
        client_assertion_type:
          type: string
          enum: ["urn:ietf:params:oauth:client-assertion-type:jwt-bearer"]
        client_assertion: { type: string, description: Signed JWT }
      required: [client_assertion_type, client_assertion]
...
paths:
...
  /internal/avatar:
    post:
      operationId: internalAvatarFetch
      tags: [internal]
      summary: S2S avatar-bytes fetch for RP backends
      security: [ { edgeMutualTLS: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/ClientAssertion"
                - type: object
                  properties: { file: { type: string } }
                  required: [file]
      responses:
        "200": { description: WebP bytes, content: { image/webp: { schema: { type: string, format: binary } } } }
        "401": { description: invalid_client, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: not_found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...